requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.26.post2
Django==1.11.13
django-celery==3.2.2
django-cors-headers==2.2.0
djangorestframework==3.8.2
drf-extensions==0.3.1
kombu==3.0.37
MySQL-python==1.2.5
pytz==2018.4
vine==1.1.4
